Updates invoice accounts module to work with contact plugin.
* Removes all contact plugin related methods.
* Updates controller to redirect to and retrieve data from plugin.
