+++ src/internal/modules/invoice-accounts/forms/add-fao.js
@@ -3,25 +3,24 @@
 const Joi = require('@hapi/joi');
 
 const { formFactory, fields } = require('shared/lib/forms/');
+const { isNull } = require('lodash');
 const urlJoin = require('url-join');
-
-const options = {
-  yes: { value: 'yes', label: 'Yes' },
-  no: { value: 'no', label: 'No' }
-};
 
 /**
  * Form to request if an FAO contact should be added to the invoice account
  *
  * @param {Object} request The Hapi request object
- * @param {Boolean}  selected value used to determine what radio option should be checked
+ * @param {Object|null} sessionContact contact data from the session
   */
-const addFaoForm = (request, selected) => {
+const addFaoForm = (request, sessionContact) => {
   const { csrfToken } = request.view;
   const { regionId, companyId } = request.params;
   const action = urlJoin('/invoice-accounts/create', regionId, companyId, 'add-fao');
-  // TODO another sub option might need to be added if selected = yes
-  const checkedOption = options[selected];
+  // contact is undefined when landing on this page for the first time,
+  // if selected 'no' previously contact === null
+  const faoRequired = (sessionContact === undefined)
+    ? sessionContact
+    : !isNull(sessionContact);
 
   const f = formFactory(action, 'POST');
 
@@ -31,21 +30,28 @@ const addFaoForm = (request, selected) => {
         message: 'Select yes if you need to add a person or department as an FAO'
       }
     },
-    choices: [options.yes, options.no],
+    choices: [
+      {
+        value: true,
+        label: 'Yes'
+      },
+      {
+        value: false,
+        label: 'No'
+      }
+    ],
     hint: 'For example, FAO Sam Burridge or FAO Accounts department'
-  }, checkedOption));
+  }, faoRequired));
   f.fields.push(fields.hidden('csrf_token', {}, csrfToken));
   f.fields.push(fields.button(null, { label: 'Continue' }));
 
   return f;
 };
 
-const addFaoFormSchema = (request) => {
-  return {
-    csrf_token: Joi.string().uuid().required(),
-    faoRequired: Joi.string().required().valid(['yes', 'no'])
-  };
-};
+const addFaoFormSchema = () => Joi.object({
+  csrf_token: Joi.string().uuid().required(),
+  faoRequired: Joi.boolean().required()
+});
 
 exports.addFaoForm = addFaoForm;
 exports.addFaoFormSchema = addFaoFormSchema;
